priv/python/requirements.txt
rembg>=2.0.0
Pillow>=10.0.0

